requirements.txt
pandas
requests
discord.py
matplotlib
openpyxl
